+++ src/factory/action/check.ts
@@ -3,11 +3,9 @@ import { ActionType } from '../actionType';
 
 export type IObject = any;
 export type IResult = any;
-// export type IPurpose = any;
 export interface IAttributes<TObject, TResult> extends ActionFactory.IAttributes<ActionType.CheckAction, TObject, TResult> {
-    // purpose: IPurpose;
 }
 /**
- * 確認アクションインターフェース
+ * 確認アクション
  */
 export type IAction<TAttributes extends IAttributes<IObject, IResult>> = ActionFactory.IAction<TAttributes>;