+++ src/factory/action/check/token.ts
@@ -1,20 +1,20 @@
 import * as ActionFactory from '../../action';
 import * as CheckActionFactory from '../check';
 
-export type IAgent = ActionFactory.IParticipant;
+export type IAgent = ActionFactory.IParticipantAsWebApplication | ActionFactory.IParticipantAsPerson;
 export interface IObject {
     token: string;
 }
 export type IResult = any;
 export type IError = any;
 /**
- * トークン確認アクション属性インターフェース
+ * トークン確認アクション属性
  */
 export interface IAttributes extends CheckActionFactory.IAttributes<IObject, IResult> {
-    object: IObject;
     agent: IAgent;
+    object: IObject;
 }
 /**
- * トークン確認アクションインターフェース
+ * トークン確認アクション
  */
 export type IAction = ActionFactory.IAction<IAttributes>;