+++ src/factory/action/consume/use/reservation.ts
@@ -1,21 +1,29 @@
+import { IParticipantAsPerson, IParticipantAsWebApplication } from '../../../action';
 import { PlaceType } from '../../../placeType';
 import { IReservation } from '../../../reservation/event';
 import * as UseActionFactory from '../use';
 
+export type IAgent = IParticipantAsPerson | IParticipantAsWebApplication;
 export type IObject = IReservation[];
 export type IResult = any;
 export interface ILocation {
     typeOf: PlaceType.Place;
+    /**
+     * 入場ゲートコード
+     */
     identifier: string;
 }
 export type IPotentialActions = any;
-
+export interface IInstrument {
+    token?: string;
+}
 export interface IAttributes extends UseActionFactory.IAttributes<IObject, IResult> {
+    agent: IAgent;
+    instrument?: IInstrument;
     location?: ILocation;
     potentialActions?: IPotentialActions;
 }
-
 /**
- * 予約使用アクションインターフェース
+ * 予約使用アクション
  */
 export type IAction = UseActionFactory.IAction<IAttributes>;