+++ src/factory/action/transfer/return/reserveTransaction.ts
@@ -1,22 +1,18 @@
 import * as COA from '@motionpicture/coa-service';
 
 import * as ActionFactory from '../../../action';
-import { ActionType } from '../../../actionType';
 import { AssetTransactionType } from '../../../assetTransactionType';
 import { ISimpleOrder } from '../../../order';
-import { OrganizationType } from '../../../organizationType';
-import { Identifier as WebAPIIdentifier } from '../../../service/webAPI';
+import { Identifier as WebAPIIdentifier, IService } from '../../../service/webAPI';
+import * as ReturnActionFactory from '../return';
 
-export interface IAgent {
-    id: string;
-    typeOf: OrganizationType.Project;
-}
+export type IAgent = ActionFactory.IParticipantAsProject;
+export type IRecipient = ActionFactory.IParticipantAsSeller;
 export type IObject4COA = COA.factory.reserve.IStateReserveArgs & {
     typeOf: 'COAReserveTransaction';
 };
 export interface IObject4Chevre {
     typeOf: AssetTransactionType.Reserve;
-    // id?: string;
     transactionNumber: string;
 }
 export type IObject<T extends WebAPIIdentifier> =
@@ -24,11 +20,14 @@ export type IObject<T extends WebAPIIdentifier> =
     IObject4Chevre;
 export type IPurpose = ISimpleOrder;
 export type IResult = any;
-export interface IAttributes<TObject, TResult> extends ActionFactory.IAttributes<ActionType.ReturnAction, TObject, TResult> {
+export type IInstrument<T extends WebAPIIdentifier> = IService<T>;
+export interface IAttributes<T extends WebAPIIdentifier> extends ReturnActionFactory.IAttributes<IObject<T>, IResult> {
     agent: IAgent;
+    instrument: IInstrument<T>;
     purpose: IPurpose;
+    recipient: IRecipient;
 }
 /**
- * 予約返却アクション
+ * 予約取引返却アクション
  */
-export type IAction<TAttributes extends IAttributes<IObject<WebAPIIdentifier>, IResult>> = ActionFactory.IAction<TAttributes>;
+export type IAction<T extends WebAPIIdentifier> = ReturnActionFactory.IAction<IAttributes<T>>;