+++ CHANGELOG.md
@@ -18,6 +18,23 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v4.236.0 - 2022-06-01
+
+### Added
+
+- 予約資産取引返却タスクを追加
+
+### Changed
+
+- 返却アクションを最適化
+- Eメール送信アクションのIPurposeを定義
+- CheckアクションのIAgentを最適化
+- 予約使用アクションを最適化
+
+### Removed
+
+- 印刷アクションを削除
+
 ## v4.235.0 - 2022-05-31
 
 ### Changed